commit
arreglos clase eventos

+++ src/main/java/com/unla/grupo6/repositories/IEventoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.unla.grupo6.entities.Evento;
 
 @Repository("eventoRepository")
-public interface IEventoRepository extends JpaRepository<Evento, Integer>{
+public interface IEventoRepository extends JpaRepository<Evento, Long>{
 	
 	//obtener los eventos en fecha y hora particular
 	List<Evento> findByfechaHora(LocalDateTime fechaHora);
